Moved from using a commandline runner to initialize data to a DummyDataInitializer that listens to application ready event

+++ src/main/java/no/ntnu/stud/idata2306_project/config/JwtUtil.java
@@ -28,7 +28,7 @@ public class JwtUtil {
    */
   public String generateToken(UserDetails userDetails) {
     final long timeNow = System.currentTimeMillis();
-    final long millisecondsInHour = 60 * 60 * 1000;
+    final long millisecondsInHour = (long) 60 * 60 * 1000;
     final long timeAfterOneHour = timeNow + millisecondsInHour;
 
     return Jwts.builder()
